==> code/03_analyses/05_raw_results.Rmd <==
---
title: "Raw Results"
author: "Erin M. Buchanan"
date: "Last Knitted `r Sys.Date()`"
output: 
  html_document:
    toc: true
    toc_float: true
    toc_depth: 3
editor_options: 
  chunk_output_type: inline
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

This document presents the raw form of the results after data coding. In our analysis plan, we indicated "For each measured variable, we will report raw numbers and percentages in each response category. Additional analyses will be exploratory; we are not conducting confirmatory tests of any *a priori* hypotheses.". This document serves to show the raw format of the coded data, as well as inform the reader about the way the data was coded.

The only corrections to the data in this document are merging the same answers together (caused by a correction in our Google Form) and entering data that was accidentally missing (e.g., missing journal information, fixes to JIF).

## Libraries

```{r}
library(rio)
library(dplyr)
library(stringi)
library(tidyr)
library(psych)
library(ggplot2)
library(maps)
library(countrycode)
library(webr)
library(patchwork)
```

## Get the Data

```{r}
# full data 
DF <- import("../../data/03_analyses/data_coding.csv")

# open access data 
openaccessDF <- import("../../data/03_analyses/OA_Coding_10_2021.xlsx") %>% select(-`...9`) %>% filter(!is.na(row_num))
```

## Article Characteristics

-   Coder instructions: To identify journal impact factors use the Thomson Reuters Journal Citation Reports (<https://jcr.clarivate.com/>). For the question about the country, check the institutional affiliation of the corresponding author. If there are multiple corresponding authors, choose the first. If no corresponding author is identified, choose the first. If there are multiple affiliations for the selected author, choose the first. For the questions about study design, examine the title, abstract, and if necessary the methods section, to establish the study characteristics.

### Publication Information

-   Publication year
    -   Derived from Scopus meta-data
-   Field
    -   Derived from Scopus meta-data

```{r}
# number of articles per year 
table(DF$Year, useNA = "ifany")

# input missing journal information 
DF$Source.title[DF$ArticleID == "9586644"] <- "Artificial Intelligence Review"
DF$Source.title[DF$ArticleID == "4.16E+06"] <- "Reading & Writing Quarterly"
DF$Source.title[DF$ArticleID == "9923056"] <- "Applied Linguistics"
DF$Source.title[DF$ArticleID == "5.84E+59"] <- "Language, Culture and Curriculum"
DF$Source.title[DF$ArticleID == "8.86E+11"] <- "International Journal of Lexicography"

# a complete list of all journals 
table(DF$Source.title, useNA = "ifany")

# total number of journals 
length(unique(DF$Source.title))

# number of journals within each split 
DF %>% 
  group_by(year_split) %>% 
  summarize(n_distinct(Source.title))
```

### Language Investigated

-   What language does the article investigate?
    -   Free text response (multiple responses possible)
    -   Additional instructions: "Multiple responses possible. If there are more than 5 languages refer to as"cross linguistic". If in doubt, refer to as "unclear". If article makes claim about all languages, refer to as "universal"

```{r}
# output of original text
table(DF$ArticleLanguage, useNA = "ifany")

# lower case to normalize
DF$ArticleLanguage <- tolower(DF$ArticleLanguage)

# normalizing
DF$ArticleLanguage <- gsub('cross linguistic', 'cross-linguistic', DF$ArticleLanguage)
DF$ArticleLanguage <- gsub("portugese", "portuguese", DF$ArticleLanguage)
DF$ArticleLanguage <- gsub("nederlandse gebarentaal", "dutch sign language", DF$ArticleLanguage)
DF$ArticleLanguage <- gsub("multilingual context 10\\+ languages|multilingualism", "cross-linguistic", DF$ArticleLanguage)
DF$ArticleLanguage <- gsub("quecha", "quechua", DF$ArticleLanguage)
DF$ArticleLanguage <- gsub('"', "", DF$ArticleLanguage)
DF$ArticleLanguage <- trimws(DF$ArticleLanguage)

# small typo fixes on original text 
table(DF$ArticleLanguage, useNA = "ifany")

# table of languages by year
table(DF$ArticleLanguage, DF$year_split, useNA = "ifany")
```

### JIF

-   Journal impact factor at year of publication
    -   What is the journal impact factor of the article at the time of publication?.
    -   Free text response (numerical)
    -   Additional instructions: "To identify journal impact factors use the Thomson Reuters Journal Citation Reports (<https://jcr.clarivate.com/>). If you have no access to the website, code as"NO ACCESS", if no JIF is available for year of publication, take first available JIF after year of publication and note the year of JIF in 2.2b below. If no JIF is available at all, code as NA."
-   Year of journal impact factor (if not = year of publication)
    -   What is the year of journal impact factor (if not = year of publication)
    -   Free text response (numerical)

**Note**: We had a separate group code JIF, as only a few had access to the database with these values.

```{r}
# import JIF coded by a person with access to the database 
JIF <- import("../../data/03_analyses/JIF_update.xlsx")

# restructure JIF
JIF_new <- JIF %>%  
  filter(JIFYear != "NA") %>% 
  mutate(JIF = as.numeric(JIF),
         JIFYear = as.numeric(JIFYear),
    year_split = ifelse(JIFYear > 2016, "After-OS", "Pre-OS"))

# describe overall 
describe(JIF_new$JIF)

# describe by year 
describeBy(as.numeric(JIF_new$JIF), JIF_new$year_split)
```

### Country

-   Which country is the corresponding author based in according to their affiliation?
    -   USA / China / UK / Germany / Japan / France / Canada / Italy / India / Spain / /Unclear /Other \*
    -   Additional instructions: "For the question about the country, check the institutional affiliation of the corresponding author. If there are multiple corresponding authors, choose the first. If no corresponding author is identified, choose the first. If there are multiple affiliations for the selected author, choose the first."

```{r}
# fix up typos
DF$CountryAuthor <- gsub("The Netherands", "The Netherlands", DF$CountryAuthor)
DF$CountryAuthor <- gsub("The Netherlands", "Netherlands", DF$CountryAuthor)
DF$CountryAuthor <- gsub("Hongkong", "Hong Kong", DF$CountryAuthor)
DF$CountryAuthor <- gsub("Finnland", "Finland", DF$CountryAuthor)
DF$CountryAuthor <- gsub("Republic of Ireland", "Ireland", DF$CountryAuthor)
DF$CountryAuthor <- gsub("Republic of South Africa", "South Africa", DF$CountryAuthor)
DF$CountryAuthor <- gsub(", Germany", "", DF$CountryAuthor)
DF$CountryAuthor <- gsub(", Poland", "", DF$CountryAuthor)
DF$CountryAuthor <- gsub("Brasil", "Brazil", DF$CountryAuthor)
DF$CountryAuthor <- gsub("NA", NA, DF$CountryAuthor)
DF$CountryAuthor <- gsub("Unclear", NA, DF$CountryAuthor)

# table of country authors 
table(DF$CountryAuthor, useNA = "ifany")

# table of authors by year
table(DF$CountryAuthor, DF$year_split, useNA = "ifany")
```

### Study type/design

-   What type of study is being reported?
    -   Multiple choice (see Table 2 for response options).
    -   Additional instructions: "Mark only one oval. Examine the title, abstract, and if necessary the methods section, to establish the study characteristics."

```{r}
# fix typo change from google sheets 
DF$StudyType[grepl("^No empirical data", DF$StudyType)] <- "No empirical data"

# table of study types 
table(DF$StudyType, useNA = "ifany")

# table of study types by year 
table(DF$StudyType, DF$year_split, useNA = "ifany")
```

### Type of empirical study

-   If the article reports empirical data, what type of empirical data study is being reported?
    -   Experimental study /
    -   Corpus study /
    -   Field study or language description /
    -   Survey or interview /
    -   Typological study /
    -   Multiple study type reported /
    -   Other

```{r}
# fix NA values and trim spaces 
DF$EmpiricalStudyData <- sub("^na$", NA, DF$EmpiricalStudyData)
DF$EmpiricalStudyData <- trimws(tolower(DF$EmpiricalStudyData))

# table of study data values
table(DF$EmpiricalStudyData, useNA = "ifany")

# table of study data by year
table(DF$EmpiricalStudyData, DF$year_split, useNA = "ifany")
```

## Preregistration

-   Definitions: "Preregistration" refers to the timestamped registration of important aspects of the study (typically hypotheses, methods, and/or analysis plan) prior to commencement of the study.
-   Coder instructions: Check specific sections where these files might be located e.g., supplementary materials, appendices, author notes, methods, and results sections. Search for "*registration" or "*registered".

### Preregistration

-   Does the article state whether or not the study (or some aspect of the study) was preregistered?
    -   Yes -- there is a preregistration /
    -   No -- there is no preregistration /
    -   Other\*

```{r}
# fix typos based on change 
DF$PreReg <- gsub(" \\(GO TO SECTION 4\\)", "", DF$PreReg)

# table of original values 
table(DF$PreReg, useNA = "ifany")

# no split table necessary 
```

### Preregistration method

-   Where does the article indicate the preregistration is located?
    -   Open Science Framework (osf.io) /
    -   AsPredicted (aspredicted.org) /
    -   Registered Report /
    -   Other\*

```{r}
# table of where pre-registered 
table(DF$PreRegWhere)

# no split table necessary 
```

### Preregistration accessible

-   Can you access and open the preregistration?
    -   Yes /
    -   No /
    -   Other\*

```{r}
# table of accessibility
table(DF$PreRegAccess)

# no split table necessary 
```

### Preregistration content

-   What aspects of the study appear to be preregistered? (select all that apply)
    -   Hypotheses /
    -   Data collection /
    -   Analysis /
    -   Other\*

```{r}
# table of aspects of pre-registration 
table(DF$PreRegAspects)

# no split table necessary 
```

## Data sharing

-   Definitions: "data" refers to recorded information that supports the analyses reported in the article. For our purposes, we differentiate between different types of data. "Raw data" refers to the recorded information in its rawest, digital form, at the level of sampling units (e.g., participants, words, utterances, trials, etc). "Processed data" refers to a derived form of the data that has undergone changes from its raw state (e.g., extraction of acoustic parameters via Praat, aggregates of responses, etc.).
-   A "data availability statement" can be as simple as a url link to a data file, or as complex as a written explanation as to why data cannot be shared.
-   Coder instructions: Check the article for a data availability statement/link. They are often located in the "supplementary material", "acknowledgements", "author notes", "methods", or "results" sections. Search the article for the text "data availab\*" (to cover "data availability" and "data available"). Search for links using "www" or "http".

### Raw data type

-   What is the nature of the raw data files?
    -   Not applicable /
    -   Text file /
    -   Audio /
    -   Video /
    -   Images /
    -   Other

```{r}
# fix unclear values 
DF$RawData <- sub("Unclear", NA, DF$RawData)
DF$RawData <- sub("Not applicable", NA, DF$RawData)

# original raw data coding
table(DF$RawData, useNA = "ifany")

# table of raw data by year
table(DF$RawData, DF$year_split, useNA = "ifany")
```

### Raw data statement

-   Does the article state whether or not raw data are available?
    -   Yes, the statement says that the raw data are available /
    -   No - raw data are not available.(GO TO SECTION 4.6) /
    -   Other\*

```{r}
DF$RawDataAvl <- sub("No, raw data are not available. \\(GO TO SECTION 4.6\\)", 
                     "No, raw data are not available.  \\(GO TO SECTION 4.6\\)", 
                     DF$RawDataAvl)

# original raw data 
table(DF$RawDataAvl, useNA = "ifany")

# table of raw data and year 
table(DF$RawDataAvl, DF$year_split, useNA = "ifany")
```

### Raw data sharing method

-   How does the statement indicate that the raw data are available?
    -   Upon request from the authors /
    -   Personal or institution website /
    -   An online, third-party repository (e.g., OSF, GitHub, FigShare etc.) /
    -   Supplementary materials hosted by the journal /
    -   Available from a third party /
    -   Unclear /
    -   Other\*

```{r}
# fix NA values 
DF$RawDataWhere <- sub("^NA$", NA, DF$RawDataWhere)

# table of where raw data is 
table(DF$RawDataWhere, useNA = "ifany")

# raw data avaliability by year 
table(DF$RawDataWhere, DF$year_split, useNA = "ifany")
```

### Raw data accessibility

-   Can you access, download, and open the raw data files without an additional step such as a sign-up or a login?
    -   Yes /
    -   No /
    -   Other\*

```{r}
# table of access 
table(DF$RawDataAccess, useNA = "ifany")

# table of access by year 
table(DF$RawDataAccess, DF$year_split, useNA = "ifany")
```

### Raw data documentation

-   Are the raw data files documented, i.e., are there meta-data that state the nature and content of individual files? For text files, is there a data dictionary / "code book" that describes the nature of individual variables?
    -   Yes /
    -   No /
    -   Unclear /
    -   Other\*

```{r}
# table of documentation 
table(DF$RawDataDocument, useNA = "ifany")

# table of documentation by year 
table(DF$RawDataDocument, DF$year_split, useNA = "ifany")
```

### Processed data statement

-   Does the article state whether or not processed data are available?
    -   Yes, the statement says that the processed data are available /
    -   No, processed data are not available. /
    -   Other\*

```{r}
# fix the google coding 
DF$ProcessData <- sub(" \\(GO TO SECTION 5\\)", "", DF$ProcessData)

# table of processed data 
table(DF$ProcessData, useNA = "ifany")

# table of processed data and year 
table(DF$ProcessData, DF$year_split, useNA = "ifany")
```

### Processed data sharing method

-   How does the statement indicate that the processed data are available?
    -   Upon request from the authors /
    -   Personal or institution website /
    -   An online, third-party repository (e.g., OSF, GitHub, FigShare etc.) /
    -   Supplementary materials hosted by the journal /
    -   Available from a third party /
    -   Unclear /
    -   Other\*

```{r}
# table of where processed data 
table(DF$ProcessDataWhere, useNA = "ifany")

# table of where processed data and year split 
table(DF$ProcessDataWhere, DF$year_split, useNA = "ifany")
```

### Processed data accessibility

-   Can you access, download, and open the processed data files without an additional step such as a sign-up or a login?
    -   Yes /
    -   No /
    -   Other\*

```{r}
# table of processed data access
table(DF$ProcessDataAccess, useNA = "ifany")

# table of processed data access by year 
table(DF$ProcessDataAccess, DF$year_split, useNA = "ifany")
```

### Processed data documentation

-   Are the processed data files documented, i.e., are there meta-data that state the nature and content of individual files? Is there a data dictionary / "code book" that describes the nature of individual variables?
    -   Yes /
    -   No /
    -   Unclear /
    -   Other\*

```{r}
# table of documentation 
table(DF$ProcessDataDocument, useNA = "ifany")

# table of documentation and year 
table(DF$ProcessDataDocument, DF$year_split, useNA = "ifany")
```

## Analysis script sharing

Definition: "Analysis scripts" refers to specification of data preparation and analysis steps in the form of highly detailed step-by-step instructions for using point-and-click software (e.g., SPSS), analysis code (e.g., R), or syntax (e.g., from SPSS). - Coder instructions: Check the article for an analysis script availability statement/link. They are often located in the "supplementary material", "acknowledgements", "author notes", "methods", or "results" sections. Search for the text "analysis script" and "analysis code". Search for links using "www" or "http".

### Analysis script availability

-   Are analysis scripts available?
    -   Yes, analysis scripts are freely available /
    -   Authors point to a third party /
    -   No, analysis scripts are not available /
    -   Other\*

```{r}
# fix typos due to google and NA 
DF$AnalysisScript <- sub(" \\(GO TO SECTION 6\\)", "", DF$AnalysisScript)
DF$AnalysisScript <- sub(".*Not applicable.*|The study did not necessitate an analysis script", "Not Applicable", DF$AnalysisScript)

# table of analysis script 
table(DF$AnalysisScript, useNA = "ifany")

# table of analysis script and year 
table(DF$AnalysisScript, DF$year_split, useNA = "ifany")
```

### Analysis script sharing method

-   How are the analysis scripts accessible?
    -   Upon request from the authors /
    -   Personal or institution website /
    -   An online, third-party repository (e.g., OSF, GitHub, FigShare etc.) /
    -   Supplementary materials hosted by the journal /
    -   Available from a third party /
    -   Unclear /
    -   Other\*

```{r}
# table of analysis script where
table(DF$AnalysisScriptWhere, useNA = "ifany")

# table of analysis script where and year 
table(DF$AnalysisScriptWhere, DF$year_split, useNA = "ifany")
```

### Analysis script accessibility

-   Can you access, download, and open the analysis scripts without an additional step such as a sign-up or a login?
    -   Yes / No / Other\*

```{r}
# table of analysis script access
table(DF$AnalysisScriptAccess, useNA = "ifany")

# table of analysis script access and year 
table(DF$AnalysisScriptAccess, DF$year_split, useNA = "ifany")
```

## Materials / Methods sharing

-   Definitions: "materials / methods" refers to any study items that would be needed to repeat the study, such as stimuli, survey instruments, and computer code/software used for data collection, presentation stimuli or running experiments (not including analysis scripts, see next section), study protocols, etc. For present purposes we do not consider supplementary data/findings (e.g., additional figures or tables) to be 'materials'.
-   Coder instructions: Check the article for a materials / methods availability statement/link. They are often located in the "supplementary material", "acknowledgements", "author notes", "methods", or "results" sections. Search for links using "www" or "http".

### Materials availability

-   Are materials or additional information about the method available?
    -   Yes, materials or additional information about the method are freely available /
    -   Materials or additional information available through a third party /
    -   No, materials or additional information about the method are not available /
    -   Other\*

```{r}
# fix google coding
DF$Materials <- sub(" \\(GO TO SECTION 7\\)", "", DF$Materials)

# table of materials 
table(DF$Materials, useNA = "ifany")

# table of materials and year 
table(DF$Materials, DF$year_split, useNA = "ifany")
```

### Materials sharing method

-   How are the materials or additional information about the method accessible?
    -   Upon request from the authors /
    -   Personal or institution website /
    -   An online, third-party repository (e.g., OSF, GitHub, FigShare etc.) /
    -   Supplementary materials hosted by the journal /
    -   Available from a third party /
    -   Unclear /
    -   Other\*
-   Additional instructions: "If authors point to a third party via a reference only, mark"Available from a third party"."

```{r}
# # table of materials where 
table(DF$MaterialsWhere, useNA = "ifany")

# table of materials where and year 
table(DF$MaterialsWhere, DF$year_split, useNA = "ifany")
```

### Materials accessibility

-   Can you access, download, and open the materials or additional information about the method without an additional step such as a sign-up or a login?
    -   Yes / No / Other\*

```{r}
# table of materials access
table(DF$MaterialsAccess, useNA = "ifany")

# table of materials access and year 
table(DF$MaterialsAccess, DF$year_split, useNA = "ifany")
```

## Replication

-   Definition: "replication" refers to repetition of a previous study's methods in order to ascertain whether similar findings can be obtained with a new sample. Coder instructions: Search the title and abstract for the phrase "replicat\*" (to cover 'replication', 'replicates', etc). Confirm the authors are using the phrase with the definition provided above.

### Replication statement

-   Does the article claim to report a replication study in abstract or title?
    -   The article claims to report a replication study (or studies) /
    -   There is no clear statement that the article reports a replication study (or studies) /
    -   Other\*

```{r}
# table of replication 
table(DF$Replication, useNA = "ifany")

# table of replication and year 
table(DF$Replication, DF$year_split, useNA = "ifany")
```

## Conflict of interest

-   Coder instructions: Conflicts of interest are usually reported in a specific section e.g., "Author information", "Conflict of interest statement", or "Acknowledgments". Search the article for the phrases "conflict of interest" and/or "competing interest".

### Conflict of Interest statement

-   Does the article include a statement indicating whether there were any conflicts of interest?
    -   Yes, the statement says that there are one or more conflicts of interest /
    -   Yes, the statement says that there is no conflict of interest /
    -   No, there is no conflict of interest statement /
    -   Other\*

```{r}
# table of COI
table(DF$COI, useNA = "ifany")

# table of COI and year 
table(DF$COI, DF$year_split, useNA = "ifany")
```

## Open access

-   Coder instructions: To establish the open access status of the article: Ensure you are not connected to a network that grants paid access to journals (e.g., a university network) or logged into a remote access system such as Shibboleth or OpenAthens. Go to <https://unpaywall.org/> and add the browser extension to your browser. Restart your browser. Enter the article's DOI URL, this should lead you to the publisher's website. Click on the unpaywall button to the right of the screen to access an open access version of the article (if the button is green, there is an OA version, if the button is grey, there is no OA version). If the article is accessible, answer "Yes". If the article is not accessible via the add-on, answer "No".

### Open access status

-   Is the article open access?
    -   Yes, is open access on the publisher's website via gold OA /
    -   Yes, is open access via unpaywall (somewhere else) /
    -   No, can only be accessed via subscription l/
    -   Other\*

```{r}
# table of open access
table(openaccessDF$open_access, useNA = "ifany")

# table of open access by year
table(openaccessDF$open_access, openaccessDF$year_split, useNA = "ifany")
```

## Export Data

```{r}
# export small typo fixes
export(DF, "../../results/03_analyses/data_typofixes.csv", row.names = F)
```
